Have literals produce `:string` rather than retrieving `<pattern-buffer>` separately.
I'm not sure why I had this implemented separately in the first place. From the history, it was introduced when converting to lx from Lexi, although it wasn't present when using Lexi. It may have been to work around a limitation of lx at the time, or it may be that I misunderstand something currently, and I'm reintroducing a bug now.