Handle a vlist of 0 items.
I'm hijacking this slightly, having it draw just a horizontal line. This is implemented as a special case, but it does feel like it fits with the node's purpose.

Surprisingly, there isn't actually a "draw a line" node! This is currently the only way to do that. It's awkward to construct, but it'll do.