Add support for an invisibility flag in the parsing front-ends.
This is populated to AST nodes, and is carried statefully during parse. For BNF dialects which provide "prose" nodes, these may be used to control the state with an extension:
```
kgt:invisible
kgt:visible
```

I don't like providing an extension, but I can't think of another reasonable way to have this user-controlled.