Remove "exit" [1/n] (from output functions).
Using `exit(EXIT_FAILURE)` in the middle of a function makes kgt hard to
be embedded as a library, because it interrupt the caller.

This patch removes them in all the generators. This is replaced by
returning the status (a boolean) to the caller. The caller will be a
good position to decide what to do with the error.

Issue:
https://github.com/katef/kgt/issues/38