Stop using exit in parser.
Using exit(EXIT_FAILURE) in the middle of a function makes kgt hard to
be embedded as a library, because it interrupts the caller.

The previous patch removed usage of exit for every generators:
https://github.com/katef/kgt/pull/40

This one removes usage of exit for every parser. The parsing_error
struct is introduced to list all the errors.

Issue:
- https://github.com/katef/kgt/issues/38
- https://github.com/ArthurSonzogni/Diagon/issues/19